feat: refactor A2A and MCP server registrations to use core package and update dependencies
