+++ packages/core/src/a2a/index.ts
@@ -1,0 +1,1 @@
+export { A2AServerRegistry } from "./registry";
