+++ packages/core/src/voltagent.ts
@@ -1,11 +1,11 @@
 import type { Logger } from "@voltagent/internal";
 import type { A2AServerDeps, A2AServerFactory, A2AServerLike } from "@voltagent/internal/a2a";
-import { A2AServerRegistry } from "@voltagent/internal/a2a";
 import type { MCPServerDeps, MCPServerFactory, MCPServerLike } from "@voltagent/internal/mcp";
-import { MCPServerRegistry } from "@voltagent/internal/mcp";
 import type { DangerouslyAllowAny } from "@voltagent/internal/types";
+import { A2AServerRegistry } from "./a2a";
 import type { Agent } from "./agent/agent";
 import { getGlobalLogger } from "./logger";
+import { MCPServerRegistry } from "./mcp";
 import { VoltAgentObservability } from "./observability/voltagent-observability";
 import { AgentRegistry } from "./registries/agent-registry";
 import type { IServerProvider, VoltAgentOptions } from "./types";
