+++ packages/internal/src/mcp/index.ts
@@ -1,2 +1,1 @@
 export * from "./types";
-export { MCPServerRegistry } from "./registry";
