+++ packages/mcp-server/package.json
@@ -7,6 +7,7 @@
     "@voltagent/internal": "^0.0.10"
   },
   "devDependencies": {
+    "@voltagent/core": "^1.0.0",
     "@types/node": "^24.2.1",
     "tsup": "^8.5.0",
     "typescript": "^5.8.2",
