+++ packages/server-core/package.json
@@ -3,9 +3,7 @@
   "description": "Framework-agnostic server core for VoltAgent",
   "version": "1.0.8",
   "dependencies": {
-    "@voltagent/a2a-server": "^1.0.0",
     "@voltagent/internal": "^0.0.10",
-    "@voltagent/mcp-server": "^1.0.0",
     "ai": "^5.0.12",
     "jsonwebtoken": "^9.0.2",
     "ws": "^8.18.1",
@@ -14,6 +12,9 @@
   },
   "devDependencies": {
     "@types/jsonwebtoken": "^9.0.10",
+    "@voltagent/a2a-server": "^1.0.0",
+    "@voltagent/core": "^1.0.0",
+    "@voltagent/mcp-server": "^1.0.0",
     "@types/node": "^24.2.1",
     "@types/ws": "^8.18.1",
     "tsup": "^8.5.0",
